pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Rye Django Rest API Starter"
authors = [{ name = "Yongky Putra", email = "yoyongdev@gmail.com" }]
dependencies = [
    "django~=4.2",
    "python-dotenv>=1.0.1",
    "djangorestframework>=3.15.1",
    "django-autoslug>=1.9.9",
    "django-countries>=7.6.1",
    "django-phonenumber-field>=7.3.0",
    "phonenumbers>=8.13.38",
    "drf-yasg>=1.21.7",
    "pillow>=10.3.0",
    "argon2-cffi>=23.1.0",
    "djoser>=2.2.3",
    "django-taggit>=5.0.1",
    "django-filter>=24.2",
    "django-celery-email>=3.0.0",
    "cloudinary>=1.40.0",
    "django-celery-beat>=2.6.0",
    "redis>=5.0.5",
    "celery>=5.4.0",
    "flower>=2.0.1",
    "psycopg2-binary>=2.9.9",
    "drf-spectacular>=0.27.2",
]
readme = "README.md"
requires-python = ">= 3.12"
license = { text = "MIT" }

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    # lint
    "ruff>=0.4.8",
    "mypy>=1.10.0",
    "pre-commit>=3.7.1",
    "bandit>=1.7.8",
    # test
    "pytest>=8.2.2",
    "pytest-django>=4.8.0",
    "pytest-cov>=5.0.0",
    "django-stubs>=5.0.2",
    # docs
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.26",
    # dev
    "django-debug-toolbar>=4.4.2",
    "watchfiles>=0.22.0",
]

[tool.rye.scripts]
start = { chain = ['migrate', 'collectstatic', 'runserver'] }
manage = { call = "manage" }
migrate = "python manage.py migrate --no-input"
collectstatic = "python manage.py collectstatic --no-input"
runserver = "python manage.py runserver localhost:8000"
lint = "pre-commit run --all-files"
docs = "mkdocs serve"
test = "pytest -v tests"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle - Error
    "W",   # pycodestyle - Warning
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "I",   # isort
    "S",   # bandit
]
ignore = [
    "F401", # unused-import
]

[tool.ruff.lint.per-file-ignores]
"**/{tests}/*" = ["S101"]

[tool.ruff.lint.isort.sections]
"django" = ["django"]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.bandit]
exclude_dirs = ["tests", "node_modules"]

[tool.mypy]
strict = true

[tool.django-stubs]
django_settings_module = "config.settings.test"


[tool.pytest.ini_options]
addopts = "--nomigrations --ds config.settings.test --cov-config=coverage.ini"

# Python warnings documentation: https://docs.python.org/3/library/warnings.html
filterwarnings = [
    # Set the default behavior for warnings to error, so it forces you to look into fixing them
    "error",
]

[tool.coverage.run]
branch = true
source = ["."]
omit = [".venv/*", "manage.py", "*migrations*"]
